+++ resources/js/components/plugins/MapGisPropertiesLabels.vue
@@ -24,11 +24,11 @@
                                 track-by="id"
                                 v-model="selectedAttribute"
                                 :allowEmpty="true"
-                                :closeOnSelect="true"
+                                :closeOnSelect="false"
                                 :customLabel="translateLabel"
                                 :disabled="useEntityName"
-                                :hideSelected="false"
-                                :multiple="false"
+                                :hideSelected="true"
+                                :multiple="true"
                                 :options="attributes"
                                 :placeholder="$t('global.select.placehoder')"
                                 :select-label="$t('global.select.select')"
@@ -365,7 +365,7 @@
                             options.getText = feature => {
                                 const eid = feature.get('entity').id;
                                 const data = this.cachedData[this.cacheKey][eid];
-                                return data ? data.value : '';
+                                return data || '';
                             };
 
                             if(this.onUpdate) {
@@ -376,14 +376,32 @@
                             }
                         } else {
                             const id = this.layer.entity_type_id;
-                            const aid = this.selectedAttribute.attribute_id;
-                            $httpQueue.add(() => $http.get(`entity/entity_type/${id}/data/${aid}?geodata=has`).then(response => {
-                                this.cachedData[this.cacheKey] = response.data;
+                            (async _ => {
+                                for(let i=0; i<this.selectedAttribute.length; i++) {
+                                    const aid = this.selectedAttribute[i].attribute_id;
+                                    await $httpQueue.add(() => $http.get(`entity/entity_type/${id}/data/${aid}?geodata=has`).then(response => {
+                                        if(!this.cachedData[this.cacheKey]) {
+                                            let data = {};
+                                            for(let k in response.data) {
+                                                data[k] = response.data[k].value;
+                                            }
+                                            this.cachedData[this.cacheKey] = data;
+                                        } else {
+                                            for(let k in response.data) {
+                                                if(!this.cachedData[this.cacheKey][k]) {
+                                                    this.cachedData[this.cacheKey][k] = response.data[k].value;
+                                                } else {
+                                                    this.cachedData[this.cacheKey][k] += `, ${response.data[k].value}`;
+                                                }
+                                            }
+                                        }
+                                    }));
+                                }
 
                                 options.getText = feature => {
                                     const eid = feature.get('entity').id;
                                     const data = this.cachedData[this.cacheKey][eid];
-                                    return data ? data.value : '';
+                                    return data || '';
                                 };
 
                                 if(this.onUpdate) {
@@ -392,7 +410,7 @@
                                         data: options
                                     });
                                 }
-                            }));
+                            })();
                         }
                     } else if(this.useEntityName) {
                         options.getText = feature => {
@@ -582,7 +600,10 @@
         },
         computed: {
             cacheKey() {
-                return `${this.layer.id}_${this.selectedAttribute.attribute_id}`;
+                const ids = this.selectedAttribute.reduce((acc, curr) => {
+                    return acc += `_${curr.attribute_id}`;
+                }, '')
+                return `${this.layer.id}_${ids}`;
             }
         }
     }