+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Chart overlays to GIS layer styling
 ### Fixed
 - Unnecessary initializing snapping to features on the map (snapping is now initialized when editing mode is activated and onyl for the current visible part of the map)
+- Drag & Drop in entity tree
 
 ## 0.7.6
 ### Added